Refactor HDF5 tests into their own file
